Modificacion Fuente

+++ database/migrations/2016_11_22_140037_tabla_presupuesto.php
@@ -18,7 +18,7 @@ class TablaPresupuesto extends Migration
             $table->string('Nombre_Actividad');
             $table->dateTime('fecha_inicio');
             $table->dateTime('fecha_fin');
-            $table->integer('presupuesto');
+            $table->bigInteger('presupuesto');
         });
     }
