+++ database/migrations/2016_11_22_193711_tabla_actividad.php
@@ -22,7 +22,7 @@ class TablaActividad extends Migration
             $table->string('Nombre');
             $table->dateTime('fecha_inicio');
             $table->dateTime('fecha_fin');
-            $table->integer('valor');
+            $table->bigInteger('valor');
             $table->string('descripcion');
 
         });
